requirements.txt
fastmcp
python-dotenv
aiomysql